feat: initial commit with the first fizzbuzz implementation
I use docker, CMake, Ninja and Catch2 to do the unit
test of this feature. The unit test tries to use the BDD
formalism.

This first implementation returns Fizz when n is a
multiple of 3, Buzz when n is a multiple of 5 and
FizzBuzz when n is a multiple of both 3 and 5.