feat: add the forth rule of bowling
If he gets a spare or strike in the last frame, the bowler gets to throw one or two more bonus balls. These bonus throws are taken as part of the same turn. If the bonus throws knock down all the pins, the process does not repeat: the bonus throws are only used to calculate the score of the final frame