feat: add the second rule of poker_hands
Pair: 2 of the 5 cards in the hand have the same value. Hands which both contain a pair are ranked by the value of the cards forming the pair. If these values are the same, the hands are ranked by the values of the cards not forming the pair, in decreasing order.

To make the PokerHand implementation easier, a sorted array of card
is used in PokerHand constructors. The highest cards are placed first.